+++ R/jit-registry.R
@@ -76,6 +76,5 @@
   list(name = "nv_upper_tri_like", static = 2:4),
   list(name = "nv_runif", static = c(1L, 3L, 4L, 5L)),
   list(name = "nv_rnorm", static = c(1L, 3L, 4L, 5L)),
-  list(name = "nv_rbinom", static = c(1L, 3L, 4L, 5L)),
-  list(name = "nv_rdunif", static = c(1L, 3L, 4L))
+  list(name = "nv_rbinom", static = c(1L, 3L, 4L, 5L))
 )